Updates value calculating logic
Previously I choose values by the progression on the base axis,
so if we were iterating on the X axis, and we were at the half of the
x values, then the y value got to be y1*0.5.
This caused scenarios where the formula y*progression produced non
continous lines, so had to switch to a different approach.'
Now we continously increase the values by the steepness of the line.