Fixes p0,p1 inaccuracies
There has been many issues with the previous solution.
I wanted to avoid too normalise the params, and ignore if the line is
going backwards or is descending, but I also wanted to avoid too many
copies. In the end I choose to swap the variables instead of the dynamic
array referencing, because it is easier to me to follow (although its slower I
imagine) with the plane transformations done in the beginning instead of
dynamically done with the flags.